chartsapi - Adds charts api post function
