+++ amplitude_python_sdk/v2/__init__.py
@@ -4,5 +4,6 @@ Contains all clients and models pertaining to the Amplitude V2 API.
 
 from .client import AmplitudeV2APIClient
 from .models.event import EventIdentifiers, EventLocationData, EventV2
+from .models.charts import ChartAnnotationsV2
 
-__all__ = ["AmplitudeV2APIClient", "EventIdentifiers", "EventLocationData", "EventV2"]
+__all__ = ["AmplitudeV2APIClient", "EventIdentifiers", "EventLocationData", "EventV2", "ChartAnnotationsV2"]
