+++ amplitude_python_sdk/v2/routes.py
@@ -2,3 +2,4 @@
 
 EVENT_API = "/httpapi"
 BATCH_API = "/batch"
+CHART_ANNOTATIONS_API = "/annotations"
